load and transform data
Now we finally have something! We can load data in an Excel file and we can then transform them to only have binary columns for RQ1. This also works by essentially looking for NA values and replacing them with zero and then saying that zero values should be replaced by zero and non-zero (like strings) should be replaced by one and added in the new columns.