viz and statistics
Now we have visualization and statistics! And there is a bunch of good stuff in here, including a JSON file generated with results. I note that there was an issue (now fixed) where the statistics and viz sliced the indeces of the dataframe wrong, meaning they would miss and ethics principle. That is also fixed here!