chance-corrected-R
We now use the chance-corrected R for effect sizes for the ethical principles and pipeline stages, respectively. This comes from the library stikpetP. Further, I have also made it so that we 1) get the df for the overall cochran's q comparisons and 2) clearly delineate the cochran's q calculation between the overall and pairwise tests. It seems like 2) makes no difference but it is better to read.